# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Signa API
  version: 0.1.0
  description: Canonical HTTP contract between the Signa Go backend and TypeScript frontend.
servers:
  - url: http://localhost:8080
    description: Local development
paths:
  /healthz:
    get:
      operationId: getHealth
      summary: Check API health
      security: []
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                status: ok
  /auth/session:
    get:
      operationId: getCurrentSession
      summary: Resolve the authenticated browser session
      description: Returns the canonical principal resolved from the HttpOnly session cookie.
      security:
        - CookieAuth: []
      responses:
        '200':
          description: Authenticated session principal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentSession'
        '401':
          description: Missing, malformed, expired, or revoked session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reports:
    post:
      operationId: createReport
      summary: Accept a text-first report
      security: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 255
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReportRequest'
      responses:
        '202':
          description: Report accepted for asynchronous processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportAcknowledgement'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many report submissions
          headers:
            Retry-After:
              schema:
                type: string
              description: Seconds to wait before retrying.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Idempotency key conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /incidents:
    get:
      operationId: listIncidents
      summary: List active public incidents
      description: Returns privacy-safe generalized geometry for active incidents only.
      security: []
      responses:
        '200':
          description: Active public incidents ordered deterministically by update time.
          content:
            application/json:
              schema:
                type: array
                maxItems: 100
                items:
                  $ref: '#/components/schemas/PublicIncident'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /incidents/{incident_id}:
    get:
      operationId: getIncident
      summary: Get one active public incident
      description: Returns privacy-safe generalized geometry without internal center points or report evidence.
      security: []
      parameters:
        - $ref: '#/components/parameters/IncidentID'
      responses:
        '200':
          description: Public incident
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicIncident'
        '400':
          description: Invalid incident identifier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Active incident not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reports/{report_id}/media/uploads:
    post:
      operationId: authorizeReportMediaUpload
      summary: Authorize a direct private media upload
      security: []
      parameters:
        - $ref: '#/components/parameters/ReportID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaUploadRequest'
      responses:
        '200':
          description: Short-lived upload authorization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaUploadAuthorization'
        '400': { description: Invalid media request, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: Report not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '429': { description: Rate limit exceeded, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '503': { description: Media storage unavailable, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /reports/{report_id}/media:
    post:
      operationId: attachReportMedia
      summary: Confirm and attach a completed direct media upload
      security: []
      parameters:
        - $ref: '#/components/parameters/ReportID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaConfirmationRequest'
      responses:
        '201':
          description: Media metadata attached to the report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportMedia'
        '400': { description: Invalid media request, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: Report not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '409': { description: Media attachment conflict, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '429': { description: Rate limit exceeded, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '503': { description: Media storage unavailable, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: signa_session
      description: HttpOnly opaque session cookie. The browser supplies it with credentialed API and SSE requests; credentials never appear in URLs.
  parameters:
    ReportID:
      name: report_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IncidentID:
      name: incident_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    CurrentSession:
      type: object
      required: [user_id]
      additionalProperties: false
      properties:
        user_id:
          type: string
          format: uuid
          description: Canonical user identity resolved by the API session store.
    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          const: ok
          description: Health status returned by the API.
    Report:
      type: object
      description: Report state returned by the backend; exact device coordinates are never returned here.
      properties:
        id:
          type: string
        reporter_id:
          type: [string, 'null']
        incident_id:
          type: [string, 'null']
        raw_text:
          type: string
        normalized_text:
          type: [string, 'null']
        source_type:
          type: [string, 'null']
        eyewitness_claim:
          type: [string, 'null']
        observed_at:
          type: [string, 'null']
          format: date-time
        submitted_at:
          type: string
          format: date-time
        language:
          type: [string, 'null']
        evidence_state:
          type: [string, 'null']
        created_at:
          type: string
          format: date-time
    CreateReportRequest:
      type: object
      required:
        - raw_text
      additionalProperties: false
      properties:
        raw_text:
          type: string
          minLength: 1
          maxLength: 10000
          description: Text of the report; surrounding whitespace is ignored for validation.
        device_location:
          anyOf:
            - $ref: '#/components/schemas/DeviceLocation'
            - type: 'null'
    DeviceLocation:
      type: object
      required:
        - latitude
        - longitude
      additionalProperties: false
      properties:
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
        accuracy:
          type: number
          minimum: 0
    ReportAcknowledgement:
      type: object
      required:
        - report_id
        - status
        - submitted_at
      properties:
        report_id:
          type: string
        status:
          type: string
          const: accepted
        submitted_at:
          type: string
          format: date-time
    MediaUploadRequest:
      type: object
      required: [media_type, content_type, size_bytes]
      additionalProperties: false
      properties:
        media_type:
          type: string
          enum: [image, audio, video]
        content_type:
          type: string
        size_bytes:
          type: integer
          format: int64
          minimum: 1
    MediaConfirmationRequest:
      type: object
      required: [object_key, media_type, content_type, size_bytes]
      additionalProperties: false
      properties:
        object_key:
          type: string
          minLength: 1
        media_type:
          type: string
          enum: [image, audio, video]
        content_type:
          type: string
        size_bytes:
          type: integer
          format: int64
          minimum: 1
    MediaUploadAuthorization:
      type: object
      required: [object_key, upload_url, required_headers, expires_at]
      properties:
        object_key: { type: string }
        upload_url: { type: string, format: uri }
        required_headers:
          type: object
          additionalProperties: { type: string }
        expires_at: { type: string, format: date-time }
    ReportMedia:
      type: object
      required: [id, report_id, object_key, media_type, content_type, size_bytes, created_at]
      properties:
        id: { type: string, format: uuid }
        report_id: { type: string, format: uuid }
        object_key: { type: string }
        media_type: { type: string, enum: [image, audio, video] }
        content_type: { type: string }
        size_bytes: { type: integer, format: int64 }
        created_at: { type: string, format: date-time }
    PublicIncident:
      type: object
      description: Active incident projection with generalized public geometry; exact internal coordinates and report evidence are omitted.
      required: [id, event_type, status, confidence_state, severity, public_geometry, started_at, last_signal_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        event_type: { type: [string, 'null'] }
        status:
          type: string
          enum: [OPEN, RESOLVING]
        confidence_state: { type: string }
        severity: { type: [string, 'null'] }
        public_geometry:
          $ref: '#/components/schemas/PublicIncidentGeometry'
        started_at: { type: [string, 'null'], format: date-time }
        last_signal_at: { type: [string, 'null'], format: date-time }
        updated_at: { type: string, format: date-time }
    PublicIncidentGeometry:
      description: GeoJSON-compatible generalized incident area in EPSG:4326.
      oneOf:
        - $ref: '#/components/schemas/PublicIncidentPolygonGeometry'
        - $ref: '#/components/schemas/PublicIncidentMultiPolygonGeometry'
    PublicIncidentPolygonGeometry:
      type: object
      required: [type, coordinates]
      properties:
        type: { type: string, const: Polygon }
        coordinates:
          description: GeoJSON positions use [longitude, latitude] ordering.
          type: array
          minItems: 1
          items:
            type: array
            minItems: 4
            items:
              type: array
              minItems: 2
              maxItems: 3
              items: { type: number }
    PublicIncidentMultiPolygonGeometry:
      type: object
      required: [type, coordinates]
      properties:
        type: { type: string, const: MultiPolygon }
        coordinates:
          description: GeoJSON positions use [longitude, latitude] ordering.
          type: array
          minItems: 1
          items:
            type: array
            minItems: 1
            items:
              type: array
              minItems: 4
              items:
                type: array
                minItems: 2
                maxItems: 3
                items: { type: number }
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
    ReportLocation:
      type: object
      description: Location associated with a report; exact retention and public precision remain policy decisions.
      properties:
        latitude:
          type: number
        longitude:
          type: number